+++ packages/oc-template-react-compiler/src/lib/hmr.ts
@@ -1,0 +1,29 @@
+import reactPlugin from '@vitejs/plugin-react';
+
+export const appBlock = ({
+  name,
+  version,
+  entry
+}: {
+  name: string;
+  version: string;
+  entry: string;
+}) => `
+<div id="root"></div>
+<script type="module">
+  const data = __INITIAL_DATA__;
+  window.__$$oc_initialData__ = Object.freeze(data);
+  window.__$$oc_Settings__ = Object.freeze({ baseUrl: '/', id: 'root', name: '${name}', staticPath: '/', version: '${version}' });
+
+  import App from "${entry}";
+  import React from 'react';
+  import { createRoot } from 'react-dom/client';
+
+  const app = React.createElement(App, data);
+  const container = document.getElementById('root');
+  const root = createRoot(container);
+  root.render(app);
+</script>
+`;
+
+export { reactPlugin as plugin };