+++ packages/oc-template-preact-compiler/CHANGELOG.md
@@ -1,5 +1,11 @@
 # oc-template-preact-compiler
 
+## 0.7.4
+
+### Patch Changes
+
+- Add HMR capabilities in oc-server. Add React support.
+
 ## 0.7.3
 
 ### Patch Changes